BackEnd/requirements.txt
# Framework y extensiones principales
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Mail==0.9.1
Flask-Limiter==2.4.0
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1

# Base de datos
SQLAlchemy==1.4.23
PyMySQL==1.0.2
mysqlclient==2.0.3
redis==3.5.3

# Utilidades
Pillow==8.3.2
python-magic==0.4.24
python-dotenv==0.19.0
bcrypt==3.2.0
pyjwt==2.1.0
gunicorn==20.1.0
werkzeug==2.0.1

# Seguridad
cryptography==3.4.8
PyJWT==2.1.0

# Procesamiento de archivos
python-magic-bin==0.4.14; sys_platform == 'win32'